Se modifico los controladores Deck y Flash
