+++ src/main/java/edu/utn/infrastructure/adapters/in/rest/controller/DeckController.java
@@ -1,13 +1,21 @@
 package edu.utn.infrastructure.adapters.in.rest.controller;
+import java.util.List;
+
 import edu.utn.application.dto.DeckDTO;
+import edu.utn.application.dto.FlashcardDTO;
 import edu.utn.application.usecase.CreateDeckUseCase;
 import edu.utn.application.usecase.DeleteDeckUseCase;
 import edu.utn.application.usecase.ModifyDeckUseCase;
 import edu.utn.application.usecase.PracticeDeckUseCase;
 import edu.utn.infrastructure.ports.in.IDeckController;
+
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
 import java.util.UUID;
+
+import edu.utn.application.usecase.AddFlashcardToDeckUseCase;
+import edu.utn.application.usecase.ListFlashcardsUseCase;
 
 @RestController
 @RequestMapping("/api/decks")
@@ -16,12 +24,16 @@ public class DeckController implements IDeckController {
     private final DeleteDeckUseCase deleteDeckUseCase;
     private final ModifyDeckUseCase modifyDeckUseCase;
     private final PracticeDeckUseCase practiceDeckUseCase;
+    private final ListFlashcardsUseCase listFlashcardsUseCase;
+    private final AddFlashcardToDeckUseCase addFlashcardToDeckUseCase;
 
-    public DeckController(CreateDeckUseCase createDeckUseCase, DeleteDeckUseCase deleteDeckUseCase, ModifyDeckUseCase modifyDeckUseCase, PracticeDeckUseCase practiceDeckUseCase) {
+    public DeckController(CreateDeckUseCase createDeckUseCase, DeleteDeckUseCase deleteDeckUseCase, ModifyDeckUseCase modifyDeckUseCase, PracticeDeckUseCase practiceDeckUseCase, ListFlashcardsUseCase listFlashcardsUseCase, AddFlashcardToDeckUseCase addFlashcardToDeckUseCase) {
         this.createDeckUseCase = createDeckUseCase;
         this.deleteDeckUseCase = deleteDeckUseCase;
         this.modifyDeckUseCase = modifyDeckUseCase;
         this.practiceDeckUseCase = practiceDeckUseCase;
+        this.listFlashcardsUseCase = listFlashcardsUseCase;
+        this.addFlashcardToDeckUseCase = addFlashcardToDeckUseCase;
     }
 
     @PostMapping
@@ -33,8 +45,8 @@ public class DeckController implements IDeckController {
 
     @PutMapping("/{deckId}")
     @Override
-    public ResponseEntity<?> modifyDeck(@PathVariable UUID deckId, @RequestBody DeckDTO deckDto) {
-        DeckDTO result = modifyDeckUseCase.execute(deckId, deckDto);
+    public ResponseEntity<?> modifyDeck(@RequestBody DeckDTO deckDto) {
+        DeckDTO result = modifyDeckUseCase.execute(deckDto);
         return ResponseEntity.ok("Se modifico el deck: " + result.getNombre() + " correctamente");
     }
 
@@ -51,4 +63,19 @@ public class DeckController implements IDeckController {
         DeckDTO result = practiceDeckUseCase.practiceDeck(deckId);
         return ResponseEntity.ok("Se inicio la practica del deck: " + result.getNombre() + " correctamente");
     }
+
+    @GetMapping("/deck/{deckId}")
+    @Override
+    public ResponseEntity<?> listFlashcards(@PathVariable UUID deckId) {
+        List<FlashcardDTO> result = listFlashcardsUseCase.execute(deckId);
+        return ResponseEntity.ok(result);
+    }
+
+    @PostMapping("/{deckId}/flashcards")
+    @Override
+    public ResponseEntity<?> addFlashcardToDeck(@PathVariable UUID deckId, @RequestBody FlashcardDTO flashcardDto) {
+        FlashcardDTO result = addFlashcardToDeckUseCase.execute(deckId, flashcardDto);
+        return ResponseEntity.ok("Se agrego la flashcard: " + result.getPregunta() + " al deck: " + deckId + " correctamente");
+    }
+
 }
